+++ src/cache/index.ts
@@ -1,0 +1,4 @@
+export * from './cache.manager.interface';
+export * from './local-cache.manager';
+export * from './ioredis-cache.manager';
+export * from './redis-cache.manager';