+++ src/clients/hosted-login/index.ts
@@ -1,0 +1,2 @@
+export * from './hosted-login.client';
+export * from './types';